tests: Move common functions to conftest.py
Move common integration test functions to conftest.py, and expand the
metrics checks for recall to the pgvector tests.
