+++ tests/integration/test_pgvector.py
@@ -1,25 +1,5 @@
 import datetime
-import json
-import os
-import random
-import string
-import sys
-import subprocess
-from subprocess import PIPE
-
-import pytest
-from pinecone import Pinecone
-
-
-def read_env_var(name):
-    value = os.environ.get(name)
-    if value is None:
-        raise Exception(f"Environment variable {name} is not set")
-    return value
-
-
-def random_string(length):
-    return "".join(random.choice(string.ascii_lowercase) for _ in range(length))
+from conftest import check_request_counts, read_env_var, random_string, spawn_vsb_inner
 
 
 def _get_index_name() -> str:
@@ -30,72 +10,11 @@ def _get_index_name() -> str:
 
 
 def spawn_vsb(workload, timeout=60, extra_args=None):
-    if extra_args is None:
-        extra_args = []
-    env = os.environ
-    env.update(
-        {"VSB__PGVECTOR_USERNAME": "postgres", "VSB__PGVECTOR_PASSWORD": "postgres"}
-    )
-    proc = subprocess.Popen(
-        [
-            "./vsb.py",
-            "--database",
-            "pgvector",
-            "--workload",
-            workload,
-            "--json",
-            "--loglevel=DEBUG",
-        ]
-        + extra_args,
-        stdout=PIPE,
-        stderr=PIPE,
-        env=env,
-        text=True,
-    )
-    try:
-        stdout, stderr = proc.communicate(timeout=timeout)
-    except subprocess.TimeoutExpired as e:
-        # Echo whatever stdout / stderr we got so far, to aid in debugging
-        if e.stdout:
-            for line in e.stdout.decode(errors="replace").splitlines():
-                print(line)
-        if e.stderr:
-            for line in e.stderr.decode(errors="replace").splitlines():
-                print(line, file=sys.stderr)
-        raise
-    # Echo subprocesses stdout & stderr to our own, so pytest can capture and
-    # report them on error.
-    print(stdout)
-    print(stderr, file=sys.stderr)
-    return proc, stdout, stderr
-
-
-def parse_stats_to_json(stdout: str) -> list(dict()):
-    """
-    Parse stdout from VSB into a list of JSON dictionaries, one for each
-    worker which reported stats.
-    """
-    # For each task type (Populate, Search, ...) we see a JSON object,
-    # so must handle multiple JSON objects in stdout.
-    stats = []
-    while stdout:
-        try:
-            stats += json.loads(stdout)
-            break
-        except json.JSONDecodeError as e:
-            stdout = stdout[e.pos :]
-    return stats
-
-
-def check_request_counts(stdout, expected: dict()):
-    stats = parse_stats_to_json(stdout)
-    by_method = {s["method"]: s for s in stats}
-    for phase, stats in expected.items():
-        assert phase in by_method, f"Missing stats for expected phase '{phase}'"
-        for stat in stats:
-            assert by_method[phase][stat] == stats[stat], (
-                f"For phase {phase} and " f"stat {stat}"
-            )
+    extra_env = {
+        "VSB__PGVECTOR_USERNAME": "postgres",
+        "VSB__PGVECTOR_PASSWORD": "postgres",
+    }
+    return spawn_vsb_inner("pgvector", workload, timeout, extra_args, extra_env)
 
 
 class TestPgvector:
@@ -109,7 +28,11 @@ class TestPgvector:
             {
                 # Populate num_requests counts batches, not individual records.
                 "Populate": {"num_requests": 1, "num_failures": 0},
-                "Search": {"num_requests": 20, "num_failures": 0},
+                "Search": {
+                    "num_requests": 20,
+                    "num_failures": 0,
+                    "recall": lambda x: len(x) == 20,
+                },
             },
         )
 
@@ -130,7 +53,11 @@ class TestPgvector:
                 # chunk will be less than the batch size (600 / 4 < 1000), then the
                 # number of requests will be equal to the number of users - i.e. 4
                 "Populate": {"num_requests": 4, "num_failures": 0},
-                "Search": {"num_requests": 20, "num_failures": 0},
+                "Search": {
+                    "num_requests": 20,
+                    "num_failures": 0,
+                    "recall": lambda x: len(x) == 20,
+                },
             },
         )
 
@@ -153,7 +80,11 @@ class TestPgvector:
                 "Populate": {"num_requests": 4, "num_failures": 0},
                 # TODO: We should only issue each search query once, but currently
                 # we perform the query once per process (2)
-                "Search": {"num_requests": 20 * 2, "num_failures": 0},
+                "Search": {
+                    "num_requests": 20 * 2,
+                    "num_failures": 0,
+                    "recall": lambda x: len(x) == 20 * 2,
+                },
             },
         )
 
@@ -182,7 +113,11 @@ class TestPgvector:
             stdout,
             {
                 # Populate num_requests counts batches, not individual records.
-                "Search": {"num_requests": 20, "num_failures": 0},
+                "Search": {
+                    "num_requests": 20,
+                    "num_failures": 0,
+                    "recall": lambda x: len(x) == 20,
+                },
             },
         )
 
@@ -198,7 +133,11 @@ class TestPgvector:
             {
                 # Populate num_requests counts batches, not individual records.
                 "Populate": {"num_requests": 10, "num_failures": 0},
-                "Search": {"num_requests": 500, "num_failures": 0},
+                "Search": {
+                    "num_requests": 500,
+                    "num_failures": 0,
+                    "recall": lambda x: len(x) == 500,
+                },
             },
         )
 
@@ -215,6 +154,10 @@ class TestPgvector:
             {
                 # Populate num_requests counts batches, not individual records.
                 "Populate": {"num_requests": 1, "num_failures": 0},
-                "Search": {"num_requests": 20, "num_failures": 0},
+                "Search": {
+                    "num_requests": 20,
+                    "num_failures": 0,
+                    "recall": lambda x: len(x) == 20,
+                },
             },
         )
